+++ Model/Configuration.php
@@ -68,9 +68,6 @@ class Configuration
     // Store configuration key for daily catalog syncing
     const FLOW_DAILY_CATALOG_SYNC = 'flowcommerce/flowconnector/daily_catalog_sync';
 
-    // Store configuration key for preloading localized catalog cache
-    const FLOW_PRELOAD_LOCALIZED_CATALOG_CACHE = 'flowcommerce/flowconnector/preload_localized_catalog_cache';
-
     // Store configuration key for overriding final prices with regular prices
     const FLOW_REGULAR_PRICING_OVERRIDE = 'flowcommerce/flowconnector/regular_pricing_override';
 
@@ -538,25 +535,6 @@ class Configuration
 
         return (bool) $this->scopeConfig->getValue(
             self::FLOW_DAILY_CATALOG_SYNC,
-            ScopeInterface::SCOPE_STORE,
-            $storeId
-        );
-    }
-
-    /**
-     * Returns true if preload localized catalog cache is enabled
-     * @param int|null $storeId
-     * @return bool
-     * @throws NoSuchEntityException
-     */
-    public function isPreloadLocalizedCatalogCacheEnabled($storeId = null)
-    {
-        if ($storeId === null) {
-            $storeId = $this->getCurrentStoreId();
-        }
-
-        return (bool) $this->scopeConfig->getValue(
-            self::FLOW_PRELOAD_LOCALIZED_CATALOG_CACHE,
             ScopeInterface::SCOPE_STORE,
             $storeId
         );