+++ Observer/FlowConnectorCheckoutRedirectObserver.php
@@ -49,10 +49,13 @@ class FlowConnectorCheckoutRedirectObserver implements ObserverInterface
      */
     public function execute(Observer $observer)
     {
+        if (!$this->configuration->isFlowEnabled()) {
+            return;
+        }
+
         $experienceCountry = $this->sessionManager->getSessionExperienceCountry();
         $experienceCurrency = $this->sessionManager->getSessionExperienceCurrency();
-        if ($this->configuration->isFlowEnabled()
-            && $this->configuration->isRedirectEnabled()
+        if ($this->configuration->isRedirectEnabled()
             && ($redirectUrl = $this->sessionManager->getCheckoutUrlWithCart($experienceCountry, $experienceCurrency))) {
             $this->actionFlag->set('', \Magento\Framework\App\Action\Action::FLAG_NO_DISPATCH, true);
             $observer->getControllerAction()->getResponse()->setRedirect($redirectUrl);